+++ parachain/primitives/ethereum/src/receipt.rs
@@ -92,8 +92,8 @@ mod tests {
 		assert_eq!(receipt.post_state_or_status, vec!(1));
 		assert_eq!(receipt.cumulative_gas_used, 414448);
 		assert_eq!(
-			receipt.bloom.0,
-			hex!("
+			receipt.bloom,
+			(&hex!("
 				042000000000000000000000800200000000000100000000000100000000000000000000
 				000000000000000000000000020000000800000000000000002000000000000000000000
 				000000080000002200000000004000100000000000000000000000000000000000000000
@@ -102,7 +102,7 @@ mod tests {
 				000000000000000000040000000000020000000001000000000000000000000000000010
 				000000020000200000102000000000000100000000000000000000000000000000000000
 				10000000
-			"),
+			")).into(),
 		);
 		assert_eq!(receipt.logs.len(), 6);
     }
